+++ rpc/types.go
@@ -31,7 +31,7 @@ type Message struct {
 	Type         uint32
 	Offset       int64
 	Data         []byte
-	Size         uint32
+	Size         int64
 	transportErr error
 
 	ID journal.OpID //Seq and ID can apparently be collapsed into one (ID)
